made code generation for widgets
